demo_20191126/requirements.txt
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask==1.0.2
SQLAlchemy==1.2.11
redis==3.3.1
psycopg2-binary==2.8.4
gunicorn==20.0.4
